% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{julia_setup}
\alias{julia_setup}
\title{julia_setup}
\usage{
julia_setup()
}
\value{
Boolean. True if correct setup, False else.
}
\description{
Checks installation of Julia and install the needed packages.
May take little time to run.Only run the first time you use Jflimo.
}
