% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.flimo_result}
\alias{print.flimo_result}
\title{print.flimo_result}
\usage{
\method{print}{flimo_result}(x, ...)
}
\arguments{
\item{x}{Object of class flimo_result from any mod/method algorithm
of the flimo package.}

\item{...}{optional args for generic method}
}
\value{
String containing most important information about argument
of class flimo_result.
}
\description{
Prints most important information about inference results.
}
