% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.flimo_result}
\alias{summary.flimo_result}
\title{summary.flimo_result}
\usage{
\method{summary}{flimo_result}(object, ...)
}
\arguments{
\item{object}{Object of class flimo_result from any mode/method algorithm of the Jflimo package.}

\item{...}{optional args for generic method summary}
}
\value{
List containing most important information about argument of class flimo_result.
}
\description{
Most important information about inference results.
}
